Change: Update CI simulation script to remove outdated code (wip)
This commit consists of removing the outdated `simulate-ci.sh` script in the `scripts/` folder. The script served its purpose of simulating a CI pipeline locally for both backend and frontend, but it has been replaced with individual test commands provided at the bottom of the `README.md` file.

Removing the script now, as it's duplicating functionality provided in the updated documentation.